+++ plugins/fluentd_telemetry_plugin/src/app.py
@@ -20,10 +20,9 @@ sys.path.append(os.getcwd())
 # pylint: disable=wrong-import-position
 import logging
 from web_service import UFMTelemetryFluentdStreamingAPI
-from streamer import \
-    UFMTelemetryStreaming,\
-    UFMTelemetryStreamingConfigParser,\
-    UFMTelemetryConstants
+from streamer import UFMTelemetryStreaming
+from streaming_config_parser import UFMTelemetryStreamingConfigParser
+from telemetry_constants import UFMTelemetryConstants
 from streaming_scheduler import StreamingScheduler
 
 # pylint: disable=no-name-in-module,import-error
@@ -46,8 +45,8 @@ def _init_logs(config_parser):
 if __name__ == '__main__':
 
     # init app config parser & load config files
-    args = ArgsParser.parse_args("UFM Telemetry Streaming to fluentd", UFMTelemetryConstants.args_list)
-    _config_parser = UFMTelemetryStreamingConfigParser(args)
+    # args = ArgsParser.parse_args("UFM Telemetry Streaming to fluentd", UFMTelemetryConstants.args_list)
+    _config_parser = UFMTelemetryStreamingConfigParser()
 
     _init_logs(_config_parser)
